requirements.txt
streamlit==1.31.1
transformers==4.39.3
torch==2.2.1
bert-extractive-summarizer==0.10.1
nltk==3.8.1
scikit-learn==1.3.2
sentencepiece==0.1.99
